# src/infrastructure/docs/swagger.yaml
openapi: 3.0.3
info:
  title: 🏭 Industrial Inventory API
  version: 1.0.0
  description: |
    ## Sistema de Inventario Industrial Minero
    
    API robusta y escalable para la gestión de inventario de instrumentos industriales mineros construida con Clean Architecture, TypeScript y PostgreSQL.
    
    ### Características Principales
    - ✅ Gestión completa de instrumentos industriales (CRUD)
    - 🔐 Autenticación JWT con roles y permisos
    - 📊 Logging estructurado con Winston
    - 🧪 Validación robusta con Zod
    - 📚 Documentación interactiva completa
    - 🏗️ Arquitectura limpia y escalable
    - 🚀 Caché optimizado por endpoint
    
    ### Tecnologías
    - **Backend**: Node.js, TypeScript, Express
    - **Base de Datos**: PostgreSQL
    - **Autenticación**: JWT (JSON Web Tokens)
    - **Validación**: Zod
    - **Logging**: Winston
    - **Documentación**: Swagger/OpenAPI
    - **Caché**: In-memory con TTL configurable
    
    ### Arquitectura
    Este proyecto sigue los principios de Clean Architecture con 4 capas bien definidas:
    - **Domain Layer**: Entidades y reglas de negocio
    - **Application Layer**: Casos de uso y DTOs
    - **Infrastructure Layer**: Implementaciones técnicas
    - **Presentation Layer**: Controladores y HTTP
    
    ### Sistema de Roles
    El sistema implementa 6 niveles de acceso con permisos específicos:
    
    1. **ADMIN** 👑
       - Acceso total al sistema (CRUD en todos los recursos)
       - Gestión de usuarios y roles
...
    
...
servers:
...
paths:
...
  /auth/register:
    post:
      summary: 👤 Register new user
      description: |
        Registra un nuevo usuario en el sistema.
        
        ### Validaciones
        - **Email**: 
          - Formato válido (example@domain.com)
          - Único en el sistema
          - Dominio válido
          - Máximo 255 caracteres
        - **Password**: 
          - Mínimo 8 caracteres
          - Al menos una mayúscula
          - Al menos una minúscula
          - Al menos un número
          - Al menos un carácter especial
          - Máximo 100 caracteres
        - **Name**: 
          - Entre 2 y 100 caracteres
          - Solo letras, espacios y acentos
          - Sin caracteres especiales
        - **Role**: 
          - ADMIN: Acceso total al sistema
          - USER: Acceso limitado a operaciones básicas
          - VIEWER: Solo lectura
        
        ### Proceso
        1. Validación de datos de entrada
        2. Verificación de email único
        3. Encriptación de contraseña (bcrypt)
        4. Creación de usuario en base de datos
        5. Generación de token JWT
        6. Registro en logs de auditoría
        
        ### Consideraciones de Seguridad
        - Las contraseñas se almacenan encriptadas con bcrypt
        - Los tokens tienen expiración configurable (24h por defecto)
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - name
                - role
              properties:
                email:
                  type: string
                  format: email
                  example: "user@example.com"
                password:
                  type: string
                  format: password
                  example: "SecurePass123!"
                name:
                  type: string
                  example: "John Doe"
                role:
                  type: string
                  enum: [ADMIN, USER, VIEWER, MANAGER, SUPERVISOR, AUDITOR]
                  example: "USER"
      responses:
        201:
          description: Usuario registrado exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Usuario registrado exitosamente"
                  data:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 1
                      email:
                        type: string
                        example: "user@example.com"
                      name:
                        type: string
                        example: "John Doe"
                      role:
                        type: string
                        example: "USER"
                      token:
                        type: string
                        example: "eyJhbGciOiJIUzI1NiIs..."
        400:
          description: Error de validación
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        409:
          description: Email ya existe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/login:
    post:
      summary: 🔑 Login user
      description: |
        Inicia sesión de usuario y genera tokens JWT.
        
        ### Proceso
        1. Validación de credenciales
        2. Generación de access token
        3. Generación de refresh token
        4. Registro en logs de auditoría
        
        ### Tokens Generados
        - **Access Token**: 24 horas de validez
        - **Refresh Token**: 7 días de validez
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: "user@example.com"
                password:
                  type: string
                  format: password
                  example: "SecurePass123!"
      responses:
        200:
          description: Login exitoso
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Login exitoso"
                  data:
                    type: object
                    properties:
                      accessToken:
                        type: string
                        example: "eyJhbGciOiJIUzI1NiIs..."
                      refreshToken:
                        type: string
                        example: "eyJhbGciOiJIUzI1NiIs..."
                      user:
                        type: object
                        properties:
                          id:
                            type: integer
                            example: 1
                          email:
                            type: string
                            example: "user@example.com"
                          name:
                            type: string
                            example: "John Doe"
                          role:
                            type: string
                            example: "USER"
        401:
          description: Credenciales inválidas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/refresh:
    post:
      summary: 🔄 Refresh access token
      description: |
        Renueva el access token usando el refresh token.
        
        ### Proceso
        1. Validación del refresh token
        2. Verificación en blacklist
        3. Generación de nuevo access token
        4. Registro en logs de auditoría
      tags:
        - Authentication
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refreshToken
              properties:
                refreshToken:
                  type: string
                  example: "eyJhbGciOiJIUzI1NiIs..."
      responses:
        200:
          description: Token renovado exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Token renovado exitosamente"
                  data:
                    type: object
                    properties:
                      accessToken:
                        type: string
                        example: "eyJhbGciOiJIUzI1NiIs..."
        401:
          description: Token inválido o expirado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /users:
    get:
      summary: 📋 List all users
      description: |
        Obtiene la lista de todos los usuarios.
        
        ### Roles Permitidos
        - ✅ **ADMIN**: Acceso completo
        - ❌ **MANAGER**: Sin acceso
        - ❌ **SUPERVISOR**: Sin acceso
        - ❌ **USER**: Sin acceso
        - ❌ **AUDITOR**: Sin acceso
        - ❌ **VIEWER**: Sin acceso
      tags:
        - Users
      security:
        - BearerAuth: []
      responses:
        '200':
          description: ✅ Lista de usuarios obtenida exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListUsersResponse'
        '401':
          description: ❌ No autorizado
    post:
      summary: ➕ Create new user
      description: |
        Crea un nuevo usuario en el sistema.
        
        ### Roles Permitidos
        - ✅ **ADMIN**: Acceso completo
        - ❌ **MANAGER**: Sin acceso
        - ❌ **SUPERVISOR**: Sin acceso
        - ❌ **USER**: Sin acceso
        - ❌ **AUDITOR**: Sin acceso
        - ❌ **VIEWER**: Sin acceso
      tags:
        - Users
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: ✅ Usuario creado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: ❌ Datos inválidos
        '401':
          description: ❌ No autorizado
        '409':
          description: ❌ Email ya registrado
  /users/{id}:
    get:
      summary: 🔍 Get user by ID
      description: |
        Obtiene los detalles de un usuario específico.
        
        ### Roles Permitidos
        - ✅ **ADMIN**: Acceso completo
        - ❌ **MANAGER**: Sin acceso
        - ❌ **SUPERVISOR**: Sin acceso
        - ❌ **USER**: Sin acceso
        - ❌ **AUDITOR**: Sin acceso
        - ❌ **VIEWER**: Sin acceso
      tags:
        - Users
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: ✅ Usuario encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: ❌ No autorizado
        '404':
          description: ❌ Usuario no encontrado
    put:
      summary: ✏️ Update user
      description: |
        Actualiza los datos de un usuario existente.
        
        ### Roles Permitidos
        - ✅ **ADMIN**: Acceso completo
        - ❌ **MANAGER**: Sin acceso
        - ❌ **SUPERVISOR**: Sin acceso
        - ❌ **USER**: Sin acceso
        - ❌ **AUDITOR**: Sin acceso
        - ❌ **VIEWER**: Sin acceso
      tags:
        - Users
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: ✅ Usuario actualizado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: ❌ Datos inválidos
        '401':
          description: ❌ No autorizado
        '404':
          description: ❌ Usuario no encontrado
    delete:
      summary: 🗑️ Delete user
      description: |
        Elimina un usuario del sistema.
        
        ### Roles Permitidos
        - ✅ **ADMIN**: Acceso completo
        - ❌ **MANAGER**: Sin acceso
        - ❌ **SUPERVISOR**: Sin acceso
        - ❌ **USER**: Sin acceso
        - ❌ **AUDITOR**: Sin acceso
        - ❌ **VIEWER**: Sin acceso
      tags:
        - Users
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: ✅ Usuario eliminado exitosamente
        '401':
          description: ❌ No autorizado
        '404':
          description: ❌ Usuario no encontrado

# --- CATEGORIES ---
  /categories:
    get:
      summary: 📋 List all categories
      description: |
        Obtiene la lista de todas las categorías.
        
        ### Roles Permitidos
        - ✅ **ADMIN**: Acceso completo
        - ✅ **MANAGER**: Acceso completo
        - ✅ **SUPERVISOR**: Acceso completo
        - ✅ **USER**: Acceso completo
        - ✅ **AUDITOR**: Acceso completo
        - ✅ **VIEWER**: Acceso completo
      tags:
        - Categories
      security:
        - BearerAuth: []
      responses:
        '200':
          description: ✅ Lista de categorías obtenida exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListCategoriesResponse'
        '401':
          description: ❌ No autorizado
    post:
      summary: ➕ Create new category
      description: |
        Crea una nueva categoría.
        
        ### Roles Permitidos
        - ✅ **ADMIN**: Acceso completo
        - ✅ **MANAGER**: Acceso completo
        - ✅ **USER**: Acceso completo
        - ❌ **SUPERVISOR**: Sin acceso
        - ❌ **AUDITOR**: Sin acceso
        - ❌ **VIEWER**: Sin acceso
      tags:
        - Categories
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCategoryRequest'
      responses:
        '201':
          description: ✅ Categoría creada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
        '400':
          description: ❌ Datos inválidos
        '401':
          description: ❌ No autorizado
        '409':
          description: ❌ Nombre ya existe
  /categories/{id}:
    get:
      summary: 🔍 Get category by ID
      description: |
        Obtiene los detalles de una categoría específica.
        
        ### Roles Permitidos
        - ✅ **ADMIN**: Acceso completo
        - ✅ **MANAGER**: Acceso completo
        - ✅ **SUPERVISOR**: Acceso completo
        - ✅ **USER**: Acceso completo
        - ✅ **AUDITOR**: Acceso completo
        - ✅ **VIEWER**: Acceso completo
      tags:
        - Categories
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: ✅ Categoría encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
        '401':
          description: ❌ No autorizado
        '404':
          description: ❌ Categoría no encontrada
    put:
      summary: ✏️ Update category
      description: |
        Actualiza los datos de una categoría existente.
        
        ### Roles Permitidos
        - ✅ **ADMIN**: Acceso completo
        - ✅ **MANAGER**: Acceso completo
        - ✅ **USER**: Acceso completo
        - ❌ **SUPERVISOR**: Sin acceso
        - ❌ **AUDITOR**: Sin acceso
        - ❌ **VIEWER**: Sin acceso
      tags:
        - Categories
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCategoryRequest'
      responses:
        '200':
          description: ✅ Categoría actualizada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
        '400':
          description: ❌ Datos inválidos
        '401':
          description: ❌ No autorizado
        '404':
          description: ❌ Categoría no encontrada
    delete:
      summary: 🗑️ Delete category
      description: |
        Elimina una categoría del sistema.
        
        ### Roles Permitidos
        - ✅ **ADMIN**: Acceso completo
        - ❌ **MANAGER**: Sin acceso
        - ❌ **SUPERVISOR**: Sin acceso
        - ❌ **USER**: Sin acceso
        - ❌ **AUDITOR**: Sin acceso
        - ❌ **VIEWER**: Sin acceso
      tags:
        - Categories
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: ✅ Categoría eliminada exitosamente
        '401':
          description: ❌ No autorizado
        '404':
          description: ❌ Categoría no encontrada

# --- SUPPLIERS ---
  /suppliers:
    get:
      summary: 📋 List all suppliers
      description: |
        Obtiene la lista de todos los proveedores.
        
        ### Roles Permitidos
        - ✅ **ADMIN**: Acceso completo
        - ✅ **MANAGER**: Acceso completo
        - ✅ **SUPERVISOR**: Acceso completo
        - ✅ **USER**: Acceso completo
        - ✅ **AUDITOR**: Acceso completo
        - ✅ **VIEWER**: Acceso completo
      tags:
        - Suppliers
      security:
        - BearerAuth: []
      responses:
        '200':
          description: ✅ Lista de proveedores obtenida exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListSuppliersResponse'
        '401':
          description: ❌ No autorizado
    post:
      summary: ➕ Create new supplier
      description: |
        Crea un nuevo proveedor.
        
        ### Roles Permitidos
        - ✅ **ADMIN**: Acceso completo
        - ✅ **MANAGER**: Acceso completo
        - ✅ **USER**: Acceso completo
        - ❌ **SUPERVISOR**: Sin acceso
        - ❌ **AUDITOR**: Sin acceso
        - ❌ **VIEWER**: Sin acceso
      tags:
        - Suppliers
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSupplierRequest'
      responses:
        '201':
          description: ✅ Proveedor creado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupplierResponse'
        '400':
          description: ❌ Datos inválidos
        '401':
          description: ❌ No autorizado
        '409':
          description: ❌ RUT/Email ya existe
  /suppliers/{id}:
    get:
      summary: 🔍 Get supplier by ID
      description: |
        Obtiene los detalles de un proveedor específico.
        
        ### Roles Permitidos
        - ✅ **ADMIN**: Acceso completo
        - ✅ **MANAGER**: Acceso completo
        - ✅ **SUPERVISOR**: Acceso completo
        - ✅ **USER**: Acceso completo
        - ✅ **AUDITOR**: Acceso completo
        - ✅ **VIEWER**: Acceso completo
      tags:
        - Suppliers
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: ✅ Proveedor encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupplierResponse'
        '401':
          description: ❌ No autorizado
        '404':
          description: ❌ Proveedor no encontrado
    put:
      summary: ✏️ Update supplier
      description: |
        Actualiza los datos de un proveedor existente.
        
        ### Roles Permitidos
        - ✅ **ADMIN**: Acceso completo
        - ✅ **MANAGER**: Acceso completo
        - ✅ **USER**: Acceso completo
        - ❌ **SUPERVISOR**: Sin acceso
        - ❌ **AUDITOR**: Sin acceso
        - ❌ **VIEWER**: Sin acceso
      tags:
        - Suppliers
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSupplierRequest'
      responses:
        '200':
          description: ✅ Proveedor actualizado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupplierResponse'
        '400':
          description: ❌ Datos inválidos
        '401':
          description: ❌ No autorizado
        '404':
          description: ❌ Proveedor no encontrado
    delete:
      summary: 🗑️ Delete supplier
      description: |
        Elimina un proveedor del sistema.
        
        ### Roles Permitidos
        - ✅ **ADMIN**: Acceso completo
        - ❌ **MANAGER**: Sin acceso
        - ❌ **SUPERVISOR**: Sin acceso
        - ❌ **USER**: Sin acceso
        - ❌ **AUDITOR**: Sin acceso
        - ❌ **VIEWER**: Sin acceso
      tags:
        - Suppliers
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: ✅ Proveedor eliminado exitosamente
        '401':
          description: ❌ No autorizado
        '404':
          description: ❌ Proveedor no encontrado

# --- LOCATIONS ---
  /locations:
    get:
      summary: 📋 List all locations
      description: |
        Obtiene la lista de todas las ubicaciones.
        
        ### Roles Permitidos
        - ✅ **ADMIN**: Acceso completo
        - ✅ **MANAGER**: Acceso completo
        - ✅ **SUPERVISOR**: Acceso completo
        - ✅ **USER**: Acceso completo
        - ✅ **AUDITOR**: Acceso completo
        - ✅ **VIEWER**: Acceso completo
      tags:
        - Locations
      security:
        - BearerAuth: []
      responses:
        '200':
          description: ✅ Lista de ubicaciones obtenida exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListLocationsResponse'
        '401':
          description: ❌ No autorizado
    post:
      summary: ➕ Create new location
      description: |
        Crea una nueva ubicación.
        
        ### Roles Permitidos
        - ✅ **ADMIN**: Acceso completo
        - ✅ **MANAGER**: Acceso completo
        - ✅ **USER**: Acceso completo
        - ❌ **SUPERVISOR**: Sin acceso
        - ❌ **AUDITOR**: Sin acceso
        - ❌ **VIEWER**: Sin acceso
      tags:
        - Locations
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLocationRequest'
      responses:
        '201':
          description: ✅ Ubicación creada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LocationResponse'
        '400':
          description: ❌ Datos inválidos
        '401':
          description: ❌ No autorizado
        '409':
          description: ❌ Código ya existe
...
  /audit-logs:
    get:
      summary: 📋 List all audit logs
      description: |
        Obtiene la lista de todos los logs de auditoría.
        
        ### Roles Permitidos
        - ✅ **ADMIN**: Acceso completo
        - ❌ **MANAGER**: Sin acceso
        - ❌ **SUPERVISOR**: Sin acceso
        - ❌ **USER**: Sin acceso
        - ✅ **AUDITOR**: Acceso completo
        - ❌ **VIEWER**: Sin acceso
      tags:
        - Audit Logs
      security:
        - BearerAuth: []
      responses:
        '200':
          description: ✅ Lista de logs obtenida exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListAuditLogsResponse'
        '401':
          description: ❌ No autorizado
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        🔐 **Autenticación JWT**
        
        Todas las rutas (excepto /auth/login y /auth/register) requieren token JWT.
        
        ### Formato
        ```
        Authorization: Bearer <token>
        ```
        
        ### Características
        - Expiración: 24 horas
        - Refresh Token: 7 días
        - Roles incluidos en payload
        - Blacklist de tokens revocados
...